% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Calculate Adjacency Matrix}
\usage{
adjacency(clusters, samples, mat)
}
\arguments{
\item{clusters}{Cluster assignments.}

\item{samples}{Samples included in subsampling draw.}

\item{mat}{Matrix to update.}
}
\value{
An adjacency matrix update reflecting the clustering of the given
subsample.
}
\description{
This utility function updates a matrix with the adjacency information of
a given cluster assignment.
}
\details{
This function speeds up serial computation of consensus matrices.
}
