% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3C.R
\name{M3C}
\alias{M3C}
\title{Monte Carlo Consensus Clustering}
\usage{
M3C(dat, maxK = 3, montecarlo = TRUE, refMethod = "reverse-pca",
  B = 100, reps = 100, distance = "euclidean", clusterAlg = "hclust",
  innerLinkage = "average", finalLinkage = "average", pItem = 0.8,
  pFeature = 1, weightsItem = NULL, weightsFeature = NULL,
  pacWindow = c(0.1, 0.9), alpha = 0.05, p.adj = NULL, CI = NULL,
  plot = FALSE, seed = NULL, cores = 1)
}
\arguments{
\item{dat}{Probe by sample omic data matrix. Data should be filtered and
normalized prior to analysis.}

\item{maxK}{Maximum cluster number to evaluate.}

\item{montecarlo}{Generate null distribution for statistical comparison? If
\code{FALSE}, function implements the traditional consensus cluster
algorithm.}

\item{refMethod}{How should null data be generated? Options include \code{
reverse-pca}, \code{cholesky}, \code{range}, and \code{permute}.}

\item{B}{Number of null datasets to generate.}

\item{reps}{Number of subsamples to draw for consensus clustering.}

\item{distance}{Distance metric for clustering. Supports all methods
available in \code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}},
as well as those implemented in the \code{bioDist} package.}

\item{clusterAlg}{Clustering algorithm to implement. Currently supports
hierarchical (\code{"hclust"}), \emph{k}-means (\code{"kmeans"}), and
\emph{k}-medoids (\code{"pam"}).}

\item{innerLinkage}{Method to use if \code{clusterAlg = "hclust"}. See \code{
\link[stats]{hclust}}.}

\item{finalLinkage}{Method to use for clustering on consensus matrix output.
Same options as for \code{innerLinkage}.}

\item{pItem}{Proportion of items to include in each subsample.}

\item{pFeature}{Proportion of features to include in each subsample.}

\item{weightsItem}{Optional vector of item weights.}

\item{weightsFeature}{Optional vector of feature weights.}

\item{pacWindow}{Lower and upper bounds for the consensus index sub-interval
over which to calculate the PAC. Must be on (0, 1).}

\item{alpha}{Significance threshold to apply to \emph{z}-statistics.}

\item{p.adj}{Optional method for \emph{p}-value adjustment. Supports all
methods available in \code{\link[stats]{p.adjust}}.}

\item{CI}{Optional confidence interval to plot around \emph{z}-statistics.
Must be on (0, 1).}

\item{plot}{Either a Boolean indicating whether to plot results in the
graphics device, or a directory to which the plot should be exported.}

\item{seed}{Optional seed for reproducibility.}

\item{cores}{How many cores should algorithm use? Generally advisable to use
as many as possible, especially with large datasets. Setting this argument
to \code{1} means function will execute in serial.}
}
\value{
A list with \code{maxK} elements. If \code{montecarlo = TRUE}, then
the first item is a results data frame with columns for cluster number \emph{
k}, observed PAC score, expected PAC score, \emph{z}-statistic, and
\emph{p}-value. Standard errors are also returned, though may be replaced by
confidence intervals if \code{CI} is non-\code{NULL}. Adjusted \emph{
p}-values are also returned if \code{p.adj} is non-\code{NULL}.

If \code{montecarlo = FALSE}, then the first item is a results data frame
with columns for cluster number \emph{k} and PAC score.

Items two through \code{maxK} are lists corresponding to the unique values of
\emph{k}, each containing the following three elements: the consensus matrix,
tree, and cluster assignments for that \emph{k} as determined by consensus
clustering.

Plots are also optionally returned or saved. If \code{montecarlo = TRUE},
then the figure depicts \emph{z}-scores for all \emph{k}; if \code{
montecarlo = FALSE}, the figure depicts PAC scores for all \emph{k}.
}
\description{
This function implements the Monte Carlo consensus clustering algorithm.
}
\details{
M3C is a hypothesis testing framework for consensus clustering. It takes an
input matrix \code{dat}, and generates \code{B} null datasets with similar
properties but no samplewise cluster structure. The consensus cluster
algorithm is then run on each simulated matrix, with PAC scores stored for
reference. The function then consensus clusters the actual input data, and
PAC scores for each cluster number \emph{k} are tested against their
theoretical null distribution.

\code{M3C} currently supports four methods for generating null datasets from
a given input matrix. \code{reverse-pca} and \code{cholesky} use different
forms of matrix decomposition to preserve genewise covariance while
scrambling samplewise covariance. The former option is faster, while the
latter is preferable when sample size exceeds gene count. When \code{
refMethod = "range"}, expression values are drawn from uniform distributions
with minima and maxima set to each gene's observed minimum and maximum. When
\code{refMethod = "permute"}, each row's values are randomly shuffled. These
latter two methods do not preserve genewise covariance, which may bias
results.

PAC stands for proportion of ambiguous clustering. To calculate the PAC for a
given cluster number \emph{k}, we first compute the consensus matrix via
consensus clustering; then generate the empirical CDF curve for the lower
triangle of that matrix; find CDF-values for the upper and lower bounds of
the PAC window; and subtract latter number from the former. Since the
consensus matrix for a perfectly stable cluster would consist of just 1s and
0s, the ideal CDF curve is flat in the middle. The goal is therefore to
minimize the PAC. See Senbabaoglu et al. (2014) for more details.
}
\examples{
mat <- matrix(rnorm(1000 * 12), nrow = 1000, ncol = 12)
m3c <- M3C(mat, maxK = 4)

}
\references{
Monti, S., Tamayo, P., Mesirov, J., & Golub, T. (2003).
\href{http://link.springer.com/article/10.1023/A:1023949509487}{Consensus
Clustering: A Resampling-Based Method for Class Discovery and Visualization
of Gene Expression Microarray Data}. \emph{Machine Learning}, \emph{52}:
91-118.

Senbabaoglu, Y., Michailidis, G. & Li, J.Z. (2014).
\href{http://www.nature.com/articles/srep06207}{Critical limitations of
consensus clustering in class discovery}. \emph{Scientific Reports}, \emph{
4}:6207.
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
