% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{PAC}
\alias{PAC}
\title{Calculate PAC}
\usage{
PAC(consensus_mats, window, plot = FALSE)
}
\arguments{
\item{consensus_mats}{A list of consensus matrices, as created by a call to
\code{\link{consensus}} or \code{\link[ConsensusClusterPlus]{
ConsensusClusterPlus}}.}

\item{window}{Lower and upper bounds for the consensus index sub-interval
over which to calculate the PAC. Must be on (0, 1).}

\item{plot}{Return plot of PAC scores by \emph{k}?}
}
\value{
A data frame with PAC scores for each value of \emph{k} in \code{
consensus_mats}. If \code{plot = TRUE}, then PAC scores by \emph{k} are
plotted.
}
\description{
This function calculates the proportion of ambiguous clustering (PAC) for
each value of \emph{k} tested via consensus clustering.
}
\details{
Consensus clustering is a method for testing the stability of cluster
membership under resampling (Monti et al., 2003). Senbabaoglu et al. (2014)
demonstrated that traditional methods for estimating optimal cluster number
fail when probes are not independent, which they rarely are in omic data. The
authors propose a new measure, the proportion of ambiguous clustering (PAC),
which represents the increase in the empirical CDF curve for each potential 
cluster number \emph{k} over a user-defined sub-interval of the consensus 
index generated by the consensus cluster algorithm. The default settings of 
\code{window = c(0.1, 0.9)} are taken from the original PAC paper, and 
generally lead to stable results.
}
\examples{
mat <- matrix(rnorm(1000 * 12), nrow = 1000, ncol = 12)
cc <- consensus(mat, maxK = 4)
pac <- PAC(cc, plot = TRUE)

}
\references{
Monti, S., Tamayo, P., Mesirov, J., & Golub, T. (2003).
\href{http://link.springer.com/article/10.1023/A:1023949509487}{Consensus
Clustering: A Resampling-Based Method for Class Discovery and Visualization
of Gene Expression Microarray Data}. \emph{Machine Learning}, \emph{52}:
91-118.

Senbabaoglu, Y., Michailidis, G. & Li, J.Z. (2014).
\href{http://www.nature.com/articles/srep06207}{Critical limitations of
consensus clustering in class discovery}. \emph{Scientific Reports}, \emph{
4}:6207.
}
