% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_pacs.R
\name{ref_pacs}
\alias{ref_pacs}
\title{Generate Reference PAC Scores}
\usage{
ref_pacs(dat, max_k = 3, ref_method = "svd", B = 100, reps = 100,
  distance = "euclidean", cluster_alg = "hclust",
  hclust_method = "average", p_item = 0.8, p_feature = 1,
  wts_item = NULL, wts_feature = NULL, pac_window = c(0.1, 0.9),
  seed = NULL, parallel = TRUE)
}
\arguments{
\item{dat}{Probe by sample omic data matrix. Data should be filtered and
normalized prior to analysis.}

\item{max_k}{Maximum cluster number to evaluate.}

\item{ref_method}{How should null data be generated? Options include \code{
svd}, \code{cholesky}, \code{range}, and \code{permute}.}

\item{B}{Number of null datasets to generate.}

\item{reps}{Number of subsamples to draw for consensus clustering.}

\item{distance}{Distance metric for clustering. Supports all methods
available in \code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}},
as well as those implemented in the \code{bioDist} package.}

\item{cluster_alg}{Clustering algorithm to implement. Currently supports
hierarchical (\code{"hclust"}), \emph{k}-means (\code{"kmeans"}), and
\emph{k}-medoids (\code{"pam"}).}

\item{hclust_method}{Method to use if \code{cluster_alg = "hclust"}. See \code{
\link[stats]{hclust}}.}

\item{p_item}{Proportion of items to include in each subsample.}

\item{p_feature}{Proportion of features to include in each subsample.}

\item{wts_item}{Optional vector of item weights.}

\item{wts_feature}{Optional vector of feature weights.}

\item{pac_window}{Lower and upper bounds for the consensus index sub-interval
over which to calculate the PAC. Must be on (0, 1).}

\item{seed}{Optional seed for reproducibility.}

\item{parallel}{If a parallel backend is loaded and available, should the 
function use it? Highly advisable if hardware permits.}
}
\value{
A matrix with \code{B} rows and \code{max_k - 1} columns containing 
null PAC scores for each cluster number \emph{k}.
}
\description{
This function computes reference PAC scores from simulated or permuted data 
based on an input matrix.
}
\details{
Suitable reference PAC scores are essential to test the magnitude and 
significance of cluster stability. This function generates \emph{B} simulated 
or permuted datasets with similar properties to \code{dat}, but with random 
sample cluster structure. The expected value of \emph{k} for these datasets 
is therefore 1, and PAC scores for each \emph{k} form a null distribution 
that tends toward normality as \emph{B} increases. 

Just as reference PAC distributions are the theoretical core of the M3C 
approach to cluster validation, \code{ref_pacs} is the computational core 
of the \code{CCtestr} package. This function can take some time to execute, 
and should ideally be run in parallel, especially with large datasets.
}
\examples{
mat <- matrix(rnorm(1000 * 12), nrow = 1000, ncol = 12)
rp <- ref_pacs(mat, ref_method = "svd")

}
