% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{Consensus Cluster Algorithm}
\usage{
consensus(dat, maxK = 3, reps = 100, distance = "euclidean",
  clusterAlg = "hclust", innerLinkage = "average", pItem = 0.8,
  pFeature = 1, weightsItem = NULL, weightsFeature = NULL, seed = NULL,
  cores = 1)
}
\arguments{
\item{dat}{Probe by sample omic data matrix. Data should be filtered and
normalized prior to analysis.}

\item{maxK}{Maximum cluster number to evaluate.}

\item{reps}{Number of subsamples to draw.}

\item{distance}{Distance metric for clustering. Supports all methods
available in \code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}},
as well as those implemented in the \code{bioDist} package.}

\item{clusterAlg}{Clustering algorithm to implement. Currently supports
hierarchical (\code{"hclust"}), \emph{k}-means (\code{"kmeans"}), and
\emph{k}-medoids (\code{"pam"}).}

\item{innerLinkage}{Method to use if \code{clusterAlg = "hclust"}. See \code{
\link[stats]{hclust}}.}

\item{pItem}{Proportion of items to include in each subsample.}

\item{pFeature}{Proportion of features to include in each subsample.}

\item{weightsItem}{Optional vector of item weights.}

\item{weightsFeature}{Optional vector of feature weights.}

\item{seed}{Optional seed for reproducibility.}

\item{cores}{How many cores should algorithm use? Generally advisable to use
as many as possible, especially with large datasets. Setting this argument
to \code{1} means function will execute in serial.}
}
\value{
A list with \code{maxK} elements, the first of which is \code{NULL}.
Elements two through \code{maxK} are consensus matrices corresponding to
cluster numbers \emph{k} = 2 through \code{maxK}.
}
\description{
This function implements the consensus cluster algorithm.
}
\details{
Consensus clustering is a resampling procedure to evaluate cluster stability.
A user-specified proportion of samples are held out on each run of the
algorithm to test how often the remaining samples do or do not cluster
together. The result is a square, symmetric consensus matrix for each value
of cluster numbers \emph{k}. Each cell of the matrix \code{mat[i, j]}
represents the proportion of all runs including samples \code{i} and \code{j}
in which the two were clustered together.
}
\examples{
mat <- matrix(rnorm(1000 * 12), nrow = 1000, ncol = 12)
cc <- consensus(mat, maxK = 4)

}
\references{
Monti, S., Tamayo, P., Mesirov, J., & Golub, T. (2003).
\href{http://link.springer.com/article/10.1023/A:1023949509487}{Consensus
Clustering: A Resampling-Based Method for Class Discovery and Visualization
of Gene Expression Microarray Data}. \emph{Machine Learning}, \emph{52}:
91-118.
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}},
\code{\link{M3C}}
}
