% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_pacs.R
\name{ref_pacs}
\alias{ref_pacs}
\title{Generate Reference PAC Scores}
\usage{
ref_pacs(dat, maxK = 3, pca = NULL, cd = NULL,
  refMethod = "reverse-pca", B = 100, reps = 100,
  distance = "euclidean", clusterAlg = "hclust", innerLinkage = "average",
  pItem = 0.8, pFeature = 1, weightsItem = NULL, weightsFeature = NULL,
  pacWindow = c(0.1, 0.9), seed = NULL, cores = 1)
}
\arguments{
\item{dat}{Probe by sample omic data matrix. Data should be filtered and
normalized prior to analysis.}

\item{maxK}{Maximum cluster number to evaluate.}

\item{pca}{Object of class \code{prcomp}, to avoid performing 
eigendecomposition of \code{dat B} times. Only relevant if \code{
refMethod = "reverse-pca"}.}

\item{cd}{Matrix representing the Cholesky decomposition of \code{dat}'s
nearest positive definite genewise covariance matrix. Only relevant if 
\code{refMethod = "cholesky"}.}

\item{refMethod}{How should null data be generated? Options include \code{
reverse-pca}, \code{cholesky}, \code{range}, and \code{permute}.}

\item{B}{Number of null datasets to generate.}

\item{reps}{Number of subsamples to draw for consensus clustering.}

\item{distance}{Distance metric for clustering. Supports all methods
available in \code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}},
as well as those implemented in the \code{bioDist} package.}

\item{clusterAlg}{Clustering algorithm to implement. Currently supports
hierarchical (\code{"hclust"}), \emph{k}-means (\code{"kmeans"}), and
\emph{k}-medoids (\code{"pam"}).}

\item{innerLinkage}{Method to use if \code{clusterAlg = "hclust"}. See \code{
\link[stats]{hclust}}.}

\item{pItem}{Proportion of items to include in each subsample.}

\item{pFeature}{Proportion of features to include in each subsample.}

\item{weightsItem}{Optional vector of item weights.}

\item{weightsFeature}{Optional vector of feature weights.}

\item{pacWindow}{Lower and upper bounds for the consensus index sub-interval
over which to calculate the PAC. Must be on (0, 1).}

\item{seed}{Optional seed for reproducibility.}

\item{cores}{How many cores should algorithm use? Generally advisable to use
as many as possible, especially with large datasets. Setting this argument
to \code{1} means function will execute in serial.}
}
\value{
A matrix with \code{B} rows and \code{maxK - 1} columns containing 
null PAC scores for each cluster number \emph{k}.
}
\description{
This function computes reference PAC scores from simulated or permuted data 
based on an input matrix.
}
\details{
Suitable reference PAC scores are essential to test the magnitude and 
significance of cluster stability. This function generates \emph{B} simulated 
or permuted datasets with similar properties to \code{dat}, but with random 
sample cluster structure. The expected value of \emph{k} for these datasets 
is therefore 1, and PAC scores for each \emph{k} form a null distribution 
that tends toward normality as \emph{B} increases. 

Just as reference PAC distributions are the theoretical core of the M3C 
approach to cluster validation, \code{ref_pacs} is the computational core 
of the \code{M3C} package. This function can take some time to execute, and 
should ideally be run in parallel, especially with large datasets.
}
\examples{
mat <- matrix(rnorm(1000 * 12), nrow = 1000, ncol = 12)
pca <- prcomp(t(mat))
rp <- ref_pacs(mat, pca = pca, refMethod = "reverse-pca")

}
